Extracted the navigation component
